Extract from keras tests folder
