+++ tests/compatibility_tests/keras_comp_tests/compatibility_save_model_test.py
@@ -16,10 +16,7 @@ import tensorflow as tf
 
 from mct_quantizers.keras.quantizers import WeightsPOTInferableQuantizer, WeightsSymmetricInferableQuantizer, \
     WeightsUniformInferableQuantizer, WeightsLUTPOTInferableQuantizer, WeightsLUTSymmetricInferableQuantizer
-from tests.keras_tests.compatibility_tests.base_compatibility_test import BaseQuantizerBuildAndSaveTest, QUANTIZER2NAME, \
-    LAYER2NAME, QUANTIZER2LAYER2ARGS
-from tests.keras_tests.test_keras_quantization_wrapper import WEIGHT, DEPTHWISE_WEIGHT
-
+from tests.compatibility_tests.keras_comp_tests.base_compatibility_test import BaseQuantizerBuildAndSaveTest
 
 class WeightsPOTQuantizerBuildAndSaveTest(BaseQuantizerBuildAndSaveTest):
